Refactor path handling logic in main.go
The code changes refactor the path handling logic in the main.go file. Instead of directly calling homeHandler and contactHandler functions, a new pathHandler function is introduced to handle different URL paths. The switch statement inside the pathHandler function determines which handler function to call based on the URL path. This change improves code organization and readability.

Additionally, the server now listens on "localhost:3000" instead of just port 3000! This will fix the issue with firewall dialogs, as the address is now localhost instead of the implicit 0.0.0.0