Add download_tailwind.sh script and update modd.conf and README.md
The commit adds a new script called `download_tailwind.sh` which is used to download the latest version of TailwindCSS. The `modd.conf` file is updated to include running the `download_tailwind.sh` script before building the application. Additionally, the README.md file is updated with instructions on how to run the application in development mode using `modd`. The README also includes information about the tools used in this project: Go, TailwindCSS, and modd.

This commit improves the development workflow by automating the process of downloading TailwindCSS and provides clear instructions for running the application in development mode.