feat: Add user insertion functionality [skip ci]
This commit adds the ability to insert a new user into the database. The `main` function now includes code that inserts a user with the name "Jon Calhoun" and email "demo2@example.com". If an error occurs during the insertion, it will be handled by panicking. After successful insertion, a message is printed indicating that the user was inserted successfully.