feat: Add token hashing to session creation
This commit adds the functionality to hash the session token when creating a new session. The `hash` method in the `SessionService` struct uses SHA256 and base64 encoding to generate a unique token hash. The hashed token is then stored in the `TokenHash` field of the `Session` struct before being inserted into the database.