feat: Add new gallery creation functionality
This commit adds the ability to create a new gallery. It introduces a new template file for the "New Gallery" page and updates the main.go file to include the necessary routes and controllers. The user can now access the "/galleries/new" route to create a new gallery by filling out a form with a title and submitting it.