feat: Add index page for galleries
This commit adds an index page for galleries, which displays a table of the user's galleries. Each gallery is listed with its ID and title, and buttons are provided to view, edit, and delete each gallery. The "Create new gallery" button at the bottom allows users to create a new gallery.

The `Galleries` struct in `controllers/galleries.go` now includes an `Index` method that retrieves the user's galleries and renders them using the `galleries/index.gohtml` template. The necessary changes have also been made in `main.go` to register the `/galleries` route.

A new template file, `galleries/index.gohtml`, has been added to define the layout of the index page. It includes HTML markup for rendering the table of galleries and buttons.

In addition, a link to "My Galleries" has been added to the navigation bar in `templates/layouts/layout-page-tailwind.gohtml`. This link is only shown if there is a logged-in user.