Add GalleryID and FileName fields to Image struct
This commit adds the GalleryID and FileName fields to the Image struct in the gallery.go file. The GalleryID field represents the ID of the gallery that the image belongs to, while the FileName field stores the name of the image file. This change allows for better organization and retrieval of images within a gallery.