feat: Add DeleteImage function to Galleries controller
This commit adds a new function called DeleteImage to the Galleries controller. This function allows users to delete an image from a gallery. If the deletion is successful, the user is redirected to the edit page of the gallery. If there is an error, a "Something went wrong" message is displayed.